DSPBRIDGE: Change address resources to void __iomem *
This patch changes address resources to void __iomem *

Ported for linux-omap-pm  bridge-2.6.29 branch.

+++ drivers/dsp/bridge/rmgr/node.c
@@ -731,7 +731,7 @@ func_cont2:
 				"0x%x\n", status);
 			}
 
-			ulGppMemBase = hostRes.dwMemBase[1];
+			ulGppMemBase = (u32)hostRes.dwMemBase[1];
 			offSet = pulValue - dynextBase;
 			ulStackSegAddr = ulGppMemBase + offSet;
 			ulStackSegVal = (u32)*((REG_UWORD32 *)
